Use FNV-1-64a for hashtable
It seems that it's pretty easy to find two randomly-generated strings
with the same hash using the 32-bit variant. When using the 64-bit variant,
it is still possible, but harder.

+++ source/common/hashtable_api.h
@@ -72,7 +72,7 @@ hashtable_status_e hashtable_destroy(hashtable_t *table);
  * @return       HASHTABLE_OK if successful, #hastable_status_e otherwise
  */
 hashtable_status_e hashtable_put(hashtable_t *table, char *key, void *data,
-                                 uint32_t *hash_output);
+                                 uint64_t *hash_output);
 
 
 /**
